requirements.txt
streamlit
chromadb
PyMuPDF
ollama
pydantic
python-docx